mobile/requirements.txt
Appium-Python-Client
assertpy
pytest
requests
selenium
pytest-html
pyotp
pytest-xdist
psutil